Refactored server utilities folder.
- Created Project router directory
- Created Algorithm router directory
- Moved router-specific util functions to dedicated _utils directories
- Renamed some files to better reflect their purpose
