+++ src/server/routers/institution/instance/algorithm/_utils/generate-args.ts
